% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agolgets.R
\name{agolGetData}
\alias{agolGetData}
\title{Fonction qui se connecte à ArcGIS online, sur les flux ASQAAs}
\usage{
agolGetData(aasqa, flux, where = "1=1", outFields = "*",
  getGeometry = TRUE)
}
\arguments{
\item{aasqa}{identifiant 'codé' de l'AASQA correspondant à ses urls ArcGIS}

\item{flux}{le nom du flux à interroger ('ind_nouvelle_aquitaine' par exemple)}

\item{where}{une requête à faire sur la table du flux. Attention les espaces
(' ') doivent être remplacés par la chaîne de caractères '%20', sinon ça
fonctionne comme une requête SQL.}

\item{outFields}{Vecteurs de chaînes de caractères indiquant les champs à
récupérer. La valeur par défaut '*' signifie 'tous'.}

\item{getGeometry}{Si TRUE, un Spatial*DataFrame est retourné, sinon une data.frame.}
}
\description{
Cette fonction permet de faire une requête sur le flux et donc de n'en 
rappatrier qu'une partie (au lieu du flux complet). Sur des gros flux,
cela permet un gain de temps significatif quand seules quelques valeurs sont
recherchées.
}
\examples{
ingeom <- agolGetData(aasqa='AQwB7zFo5jNpvM0X', flux='ind_nouvelle_aquitaine',
                       where="date\%20IN\%20('2018-09-17','2018-09-18','2018-09-19')")
indice <- agolGetData(aasqa='AQwB7zFo5jNpvM0X', flux='ind_nouvelle_aquitaine',
                       outFields=c('lib_zone', 'date', 'valeur'),
                       where="date\%20IN\%20('2018-09-17','2018-09-18','2018-09-19')",
                       getGeometry=FALSE)
}
