% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agolgets.R
\name{agolListFields}
\alias{agolListFields}
\title{Fonction pour lister les champs d'un flux}
\usage{
agolListFields(aasqa, flux)
}
\arguments{
\item{aasqa}{identifiant 'codé' de l'AASQA correspondant à ses urls ArcGIS}

\item{flux}{le nom du flux à interroger ('ind_nouvelle_aquitaine' par exemple)}
}
\description{
Fonction pour lister les champs d'un flux
}
\examples{
agolListFields(aasqa='AQwB7zFo5jNpvM0X', flux='ind_nouvelle_aquitaine')
}
