% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses2XR.R
\name{analyses2XR}
\alias{analyses2XR}
\title{Creation d'un fichier d'import de données manuelles pour XR}
\usage{
analyses2XR(
  fichier,
  nreseau,
  fichier_export = NULL,
  sheet = 1,
  startRow = 1,
  infLQ = c("\\\\*", "< "),
  unites = "microg/m3"
)
}
\arguments{
\item{fichier}{nom du fichier d'entrée}

\item{nreseau}{numero du réseau (tel qu'il doit être intégré dans XR).}

\item{fichier_export}{nom du fichier de sortie (par défaut =
fichier+'export.csv')}

\item{sheet}{numéro ou titre de l'onglet à lire}

\item{startRow}{numéro de ligne contenant les titres de colonne (et donc
première ligne à lire dans l'onglet)}

\item{infLQ}{caractères utilisés dans le fichier d'entrée pour indiquer
que la concentration est inférieure à la limite de quantification.}

\item{unites}{chaîne de caractères à utiliser pour les unités du polluant.}
}
\description{
Cette fonction lit un fichier xlsx contenant des colonnees spécifiques et
créé un fichier importable dans XR. Le fichier d'origine doit contenir 
les colonnes suivantes :
NSIT (numéro du site dans XR), NUM_ECH (identifiant de l'échantillon à
renseigner selon les règles de l'AASQA), BT (le prélèvement est-il un
blanc ? 2 si oui, 0 si non), CODE_PREL (code de la méthode de prélèvement
dans XR), DEBUT (date de début au format date dans le fichier excel - si 
chaîne de caractères, ça ne fonctionnera pas), HDEBUT (heure de début, au
format heure), FIN et HFIN (date et heure de fin, aux formats appropriés), 
ANALYSE (date d'analyse au format date). Les colonnes HDEBUT et HFIN ne sont
pas obligatoires. LABO Nom du laboratoire ayant réalisé les analyses
('Micropolluants Technologie', 'IANESCO CHIMIE', ...).
}
\details{
En plus de ces colonnes, une colonne supplémentaire doit être ajoutée pour 
chaque polluant ayant fait l'objet d'une analyse. Le nom des colonnes doit 
être constitué de la chaîne "ISO_" suivi du code ISO du polluant en question.
Par exemple ISO_01 pour le SO2.
}
