% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reglementation.R
\name{comparaison}
\alias{comparaison}
\title{Compare un jeu de donnees a des valeurs réglementaires}
\usage{
comparaison(x, seuil, detail, check.720 = TRUE, ...)
}
\arguments{
\item{x}{jeu de données (\code{\link[timetools]{TimeIntervalDataFrame-class}})
contenant les données à mettre en forme}

\item{seuil}{seuil (cf \code{\link{seuils}}) pour lequel les données
doivent être mise en forme}

\item{detail}{booléen indiquant si le détail doit être retourné ou non}

\item{check.720}{booléen indiquant si, lorsque cela est pertinent,
    la règle des \sQuote{720 heures consécutives} doit être vérifiée
    ou non (cette règle apparaît dans un guide \sQuote{ADEME} mais
dans aucun texte réglementaire).}

\item{...}{arguments supplémentaires pouvant être transmis aux 
    fonctions appelées dans le cadre du calcul. Le plus souvent, 
    il peut s'agir de l'argument \code{use.marges} (TRUE ou FALSE)
    qui précise si le calcul doit être réalisé en prenant comme 
    seuil sa valeur définitive, ou bien cette valeur \sQuote{réhaussée}
de la marge de dépassement applicable (lorsque cela est pertinent).}

\item{\dots}{arguments supplémentaires éventuellement
utilisés lorsqu'une fonction spécifique est utilisée
pour la comparaison avec le seuil (cf \code{\link{validation.reglementaire}}
pour l'explication du mécanisme de calcul).}
}
\value{
un objet de classe \code{\link[timetools:TimeIntervalDataFrame-class]{TimeIntervalDataFrame}}
qui contient le résultat de la comparaison. Si le détail est demandé
la valeur (si le seuil consiste simplement en une valeur à ne pas
dépasser) ou le nombre de dépassement du seuil (s'il
le seuil ne doit pas être dépassé plus d'un nombre de fois donné)
est retourné. TRUE (critère dépassé) ou FALSE (critère respecté) si le détail n'est pas demandé.
}
\description{
Cette fonction prend un jeu de données au format
\code{\link[timetools:TimeIntervalDataFrame-class]{TimeIntervalDataFrame}}
et le compare a un seuil réglementaire.
}
\details{
La fonction ne fait aucune autre action que comparer
les données au seuil : aucune vérification sur le bon
formatage des données n'est réalisées (base horaire, annuelle, etc.)
Pour cela, il est possible de préparer les données au moyen de
la fonction \code{\link{preparation.base}}.
C'est ce que réalise la fonction \code{\link{validation.reglementaire}}.
Se reporter à cette dernière pour voir l'articulation de
la fonction \code{\link{preparation.base}} avec la 
fonction \code{\link{comparaison}}.
}
\seealso{
\code{\link[timetools]{TimeIntervalDataFrame-class}},
\code{\link{validation.reglementaire}}, \code{\link{seuils}},
\code{\link{preparation.base}}
}

