% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosepol.R
\name{compute.rose}
\alias{compute.rose}
\title{Calcul de roses de concentrations}
\usage{
compute.rose(values, dv, fv, breaks = 0:17 * 22.5 + 22.5/2, FUN, ...)
}
\arguments{
\item{values}{vecteur contenant les valeurs de la variable à agréger
en fonction du vent.}

\item{dv}{vecteur \sQuote{numeric} contenant les directions du vent en degrés
(0 -> Nord, 90 -> Ouest, etc.).}

\item{fv}{vecteur \sQuote{numeric} contenant la vitesse du vent en m/s.
Les valeurs de \sQuote{values} pour lesquelles \sQuote{fv} est 0 ou NA
ne sont pas prises en compte.
(Attention : la prise en compte de ce paramètre risque d'évoluer à l'avenir).}

\item{breaks}{vecteur \sQuote{numeric} indiquant les limites entre chaque secteur
    de vents (en degrés). Si une valeur indiquée est en-dehors de [0-360], 
elle est prise en compte, mais modulo 360.

Les valeurs par défaut correspondent aux 16 points cardinaux (N, NNE, NE, 
ENE, E, ESE, SE, SSE, S, etc.).}

\item{FUN}{fonction à utiliser pour l'agrégation par secteur de vent (sous forme
de \sQuote{character} ou directement la fonction).}

\item{\dots}{arguments supplémentaires pour la fonction FUN.}
}
\value{
une liste
\item{breaks}{un vecteur \sQuote{numeric} contenant les limites des secteurs de vents.
	Voir la section\sQuote{Details} pour plus de précisions.}
\item{values}{Vecteur \sQuote{numeric} de longueur égale au nombre de secteurs de vent
	 définis.}
\item{opened}{Un facteur indiquant si les secteurs de vent sont ouverts à gauche ou à droite
	(pour l'instant, seuls les secteurs ouverts à gauche sont pris en compte).}
}
\description{
Cette fonction permet de calculer des "roses de concentrations" (moyennes, maximales, etc.).
Elle permet
de répartir les valeurs prises par une variable quelconque
en fonction des directions de vents concomitantes (en tenant
compte des vents nuls) et d'appliquer une fonction d'agrégation
à chaque groupe ainsi obtenu.
}
\details{
Le résultat de cette fonction peut être directement utilsé par la 
fonction \code{\link{plot.rose}} pour afficher graphiquement 
la rose des concentrations. La fonction \code{\link{rosepol}}
est un raccourci pour l'emploi successif de ces deux fonctions.

Les trois vecteurs \sQuote{values}, \sQuote{dv}, \sQuote{fv} doivent avoir le 
même nombre d'éléments.

La fonction est construite de telle façon que l'ensemble des directions de vents
est pris en compte. Autrement dit, si \code{breaks <- c(0, 90)}, deux secteurs 
de vents sont définis : ]0 - 90] et ]90 - 360]
(la taille des deux secteurs n'est donc pas identique). D'autre part, 
deux secteurs de vents ne peuvent pas se chevaucher : si \code{breaks <- c(0, 40, 20)},
il sera modifié en \code{c(0, 20, 40)}.

Les secteurs de vent indiqués par \sQuote{breaks} sont considérés comme ouverts
à gauche. Autrement dit, dans l'exemple précédent, une valeur correspondant à un
vent de direction 0° sera rangée dans le secteur ]90 - 360] ;
et une valeur correspondant à un vent de direction 1° sera rangée dans le 
secteur ]0 - 90].

Le vecteur \sQuote{breaks} de la liste de retour diffère de celui entré en argument
si ce dernier possédait des doublons, n'était pas ordonné ou contenait des
valeurs en-dehors du domaine [0-360].

Le vecteur \sQuote{breaks} à une longueur égale au nombre de secteur plus un.
En effet, sa première valeur et sa dernière valeur sont égales à un modulo 360
près. Ce format permet d'accéder plus simplement aux extrémités de chaque secteur.
Ainsi, le secteur \sQuote{i} est délimité par ]breaks[i] - breaks[i+1]].
La première valeur de \sQuote{breaks} sera donc systématiquement nulle ou négative.
}
\examples{
compute.rose (sample (10, 50, TRUE),
	 sample (360, 50, TRUE),
	 sample (0:9, 50, TRUE, c(1, rep(10, 9))))

}
\seealso{
\code{\link{plot.rose}}, \code{\link{rosepol}}, \code{\link{compute.rosevent}}, \code{\link{rosevent}}
}
