% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rosevent.R
\name{compute.rosevent}
\alias{compute.rosevent}
\title{Calcul de roses des vents}
\usage{
compute.rosevent(dv, fv, breaks = 0:17 * 22.5 + 22.5/2, vlims)
}
\arguments{
\item{dv}{vecteur \sQuote{numeric} contenant les directions du vent en degrés
(0 -> Nord, 90 -> Ouest, etc.).}

\item{fv}{vecteur \sQuote{numeric} contenant la vitesse du vent en m/s.
Les valeurs de \sQuote{values} pour lesquelles \sQuote{fv} est 0 ou NA
ne sont pas prises en compte.
(Attention : la prise en compte de ce paramètre risque d'évoluer à l'avenir).}

\item{breaks}{vecteur \sQuote{numeric} indiquant les limites entre chaque secteur
    de vents (en degrés). Si une valeur indiquée est en-dehors de [0-360],
    elle est prise en compte, mais modulo 360.

    Les valeurs par défaut correspondent aux 16 points cardinaux (N, NNE, NE,
    ENE, E, ESE, SE, SSE, S, etc.).}

\item{vlims}{vecteurs contenant les limites des différentes classes de vitesses
    de vent. Voir la section \sQuote{Details} pour la prise en compte des vents
    calmes.}
}
\value{
une liste de listes structurées de manière identique à ce qui
	est obtenu en utilisant la fonction \code{\link{compute.rose}} ainsi
	qu'un élément supplémentaire : \code{vlim}.
	Chaque élément de la liste contient les éléments suivants :
\item{breaks}{un vecteur \sQuote{numeric} contenant les limites des secteurs de vents.
		Voir la section\sQuote{Details} pour plus de précisions.}
\item{values}{Vecteur \sQuote{numeric} de longueur égale au nombre de secteurs de vent
		 définis.}
\item{opened}{Un facteur indiquant si les secteurs de vent sont ouverts à gauche ou à droite
		(pour l'instant, seuls les secteurs ouverts à gauche sont pris en compte).}
\item{vlim}{Valeur numérique correspondant à la vitesse de vent à laquelle correspond
		l'élément en question (\code{0} correspond aux vents nuls et calmes).}
}
\description{
Cette fonction permet de calculer des "roses des vents".
}
\details{
Le résultat de cette fonction peut être directement utilsé par la
fonction \code{\link{plot.rosevent}} pour afficher graphiquement
la rose des vents. La fonction \code{\link{rosevent}}
est un raccourci pour l'emploi successif de ces deux fonctions.

Les deux vecteurs \sQuote{dv}, \sQuote{fv} doivent avoir le
même nombre d'éléments.

La fonction \code{compute.rosevent} utilise la fonction \code{\link{compute.rose}}
pour faire les calculs à proprement parler. Pour avoir des détails
sur l'utilisation/la définition des paramètres \code{dv}, \code{fv} et
\code{breaks} il est possible de se reporter à la-dite fonction.

La notion de \sQuote{vent calme} est prise en compte dans le calcul des roses
des vents. Pour une définition exacte, se référer à la section \sQuote{References}.
D'une manière simplifiée, les vents calmes sont des vents dont la vitesse
est trop faible pour considérer qu'un régime de vent est établi.
Dans ce cas, le vent n'est pas comptabilisé comme venant d'une direction particulière,
il est traité comme un vent nul.

Les vents calmes sont définis à l'aide du paramètre \code{vlims}.
\itemize{
\item si la valeur \code{0} est contenue dans \code{vlims}, alors
	les vents calmes ne sont pa pris en compte et seuls les vents
	nuls sont traités à part ;
\item si la valeur \code{0} n'est pas dans \code{vlims}, alors
	les vents calmes sont définis comme les vents ayant une vitesse
	inférieure ou égale à la plus faible valeur spécifiée dans \code{vlims}.
}
}
\references{
\href{http://comprendre.meteofrance.com/jsp/site/Portal.jsp?&page_id=2866&document_id=1195&portlet_id=1796}{Meteo France}
}
\seealso{
\code{\link{plot.rosevent}}, \code{\link{rosevent}}, \code{\link{compute.rose}}, \code{\link{rosepol}}
}

