% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reglementation.R
\name{preparation.base}
\alias{preparation.base}
\title{Mise en forme des donnees avant calculs reglementaires}
\usage{
preparation.base(x, seuil, base = c("calcul", "comparaison"),
  check.720 = TRUE, ...)
}
\arguments{
\item{x}{jeu de données (\code{\link[timetools]{TimeIntervalDataFrame-class}})
contenant les données à mettre en forme}

\item{seuil}{seuil (cf \code{\link{seuils}}) pour lequel les données
doivent être mise en forme}

\item{base}{chaîne de caractères indiquant si la préparation
    est celle de l'étape \sQuote{calcul} ou \sQuote{comparaison}.
    Pour chaque seuil, les paramètres pouvant être différents entre
ces deux étapes, il est important de bien préciser celle
qui doit être réalisée (cf \code{\link{validation.reglementaire}}
    pour l'articulation des ces deux étapes).}

\item{check.720}{booléen indiquant si, lorsque cela est pertinent,
    la règle des \sQuote{720 heures consécutives} doit être vérifiée
    ou non (cette règle apparaît dans un guide \sQuote{ADEME} mais
dans aucun texte réglementaire).}

\item{\dots}{arguments supplémentaires pouvant être transmis aux 
    fonctions appelées dans le cadre du calcul. Le plus souvent, 
    il peut s'agir de l'argument \code{use.marges} (TRUE ou FALSE)
    qui précise si le calcul doit être réalisé en prenant comme 
    seuil sa valeur définitive, ou bien cette valeur \sQuote{réhaussée}
de la marge de dépassement applicable (lorsque cela est pertinent).}
}
\value{
un objet de classe \code{\link[timetools:TimeIntervalDataFrame-class]{TimeIntervalDataFrame}}
qui contient les valeurs préparées en tenant des critères de validité du seuil.
Si les critères ne sont pas respectés, les valeurs sont NAnifiées.
}
\description{
Cette fonction permet de s'assurer que les données utilisées
au moment de l'application des calculs réglementaires sont
au bon format.
}
\details{
Suivant les seuils, il peut y avoir une ou deux "pré"-paration(s)
à réaliser avant calcul, l'approche est cependant identique 
dans les cas et une seule fonction est nécessaire.
Voir \code{\link{validation.reglementaire}} pour l'explication
détaillée de l'articulation de ces deux pré-calculs avant
application des seuils.
}
\seealso{
\code{\link[timetools]{TimeIntervalDataFrame-class}},
\code{\link{validation.reglementaire}}, \code{\link{seuils}},
\code{\link{comparaison}}
}

