\name{print.seuil}
\alias{format.seuil}
\alias{print.seuil}
\title{Affichage des objets de classe seuil (classe S3)}
\usage{
  \method{print}{seuil} (x, ...)

  \method{format}{seuil} (x, ...)
}
\arguments{
  \item{x}{objet à afficher}

  \item{\dots}{pour les autres méthodes}
}
\value{
  une chaîne de caractères décrivant l'objet \code{x}
}
\description{
  Affichage des objets de classe seuil (classe S3)

  Mise forme 'lisible' des objets de classe seuil (classe
  S3)
}

