% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{qh}
\alias{d}
\alias{h}
\alias{m}
\alias{qh}
\alias{y}
\title{definition d'un raccourci pour les heures et les quart-d'heures}
\format{\preformatted{Formal class 'POSIXctp' [package "timetools"] with 2 slots
  ..@ duration: int 15
  ..@ unit    : Ord.factor w/ 8 levels "second"<"minute"<..: 2
}}
\usage{
qh

h

d

m

y
}
\description{
Pour faciliter la manipulation des heures et des quart-heures
et permettre une utilisation comme c'est le cas pour les mois
les années etc. dans \code{\link[lubridate:seconds]{lubridate}}.
}
\seealso{
\code{\link[lubridate]{seconds}}
}
\keyword{datasets}

