% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosepol.R
\name{rosepol}
\alias{rosepol}
\title{Calcul et affichage de roses de concentrations}
\usage{
rosepol(values, dv, fv, breaks = 0:17 * 22.5 + 22.5/2, FUN = "mean",
  col, border, density = -1, rlim = NULL, at.rho = NULL,
  unite = "", labels.rho = NULL, col.rho, lwd.rho = 1, lty.rho = 2,
  expr.rho = FALSE, cex.rho = 1, theta.rho = sample(360, 1),
  at.theta = 0:7 * 45, labels.theta = c("N", "NE", "E", "SE", "S",
  "SO", "O", "NO"), nb.theta = 360, col.theta, lwd.theta = 1,
  lty.theta = 2, centre = list(x = 0, y = 0), pas.rose = 1,
  facteur = 1, add = FALSE, ...)
}
\arguments{
\item{values}{vecteur contenant les valeurs de la variable à agréger
en fonction du vent.}

\item{dv}{vecteur \sQuote{numeric} contenant les directions du vent en degrés
(0 -> Nord, 90 -> Ouest, etc.).}

\item{fv}{vecteur \sQuote{numeric} contenant la vitesse du vent en m/s.
Les valeurs de \sQuote{values} pour lesquelles \sQuote{fv} est 0 ou NA
ne sont pas prises en compte.
(Attention : la prise en compte de ce paramètre risque d'évoluer à l'avenir).}

\item{breaks}{vecteur \sQuote{numeric} indiquant les limites entre chaque secteur
    de vents (en degrés). Si une valeur indiquée est en-dehors de [0-360], 
elle est prise en compte, mais modulo 360.

Les valeurs par défaut correspondent aux 16 points cardinaux (N, NNE, NE, 
ENE, E, ESE, SE, SSE, S, etc.).}

\item{FUN}{fonction à utiliser pour l'agrégation par secteur de vent (sous forme
de \sQuote{character} ou directement la fonction).}

\item{col}{couleur de remplissage de la rose.}

\item{border}{couleur de la bordure de la rose.}

\item{density}{the density of shading lines, in lines per inch.  The default
    value of ‘NULL’ means that no shading lines are drawn.  A
    zero value of ‘density’ means no shading nor filling whereas
negative values (and ‘NA’) suppress shading (and so allow
    color filling).}

\item{rlim}{limites du graphiques. contient 2 valeurs, la première est la valeur
correspondant la concentration au centre de la rose ; la seconde à celle 
au bord extérieur de la rose.}

\item{at.rho}{valeurs de rayon auxquelles doivent être affichés des divisions
pour l'échelle des concentrations.}

\item{unite}{chaîne de caractères qui sera ajouté à chaque étiquette de l'échelle
des concentrations.}

\item{labels.rho}{chaînes de caractères ou expressions (cf \sQuote{Details})
indiquant les étiquettes de l'échelle de concentrations.}

\item{col.rho}{couleur à appliquer à l'échelle des concentrations (cf \sQuote{Details}).}

\item{lwd.rho}{épaisseur de l'échelle des concentrations (cf \sQuote{Details}).}

\item{lty.rho}{type de l'échelle des concentrations (pointillés, continu) (cf \sQuote{Details}).}

\item{expr.rho}{logical indiquant si labels.rho contient des chaînes de caractères
ou des expressions (cf \sQuote{Details}).}

\item{cex.rho}{character expansion pour la graduation}

\item{theta.rho}{angle en degré où affiché les graduations}

\item{at.theta}{indique pour quels degrés les étiquettes pour l'échelle
des directions de vent doivent être affichées.}

\item{labels.theta}{chaînes de caractères contenant les étiquettes pour l'échelle
des directions de vent.}

\item{nb.theta}{indique (en degré) la précision des cercles de l'échelle des concentrations.
Un pas de 1 signifie qu'un point sera tracé tous les degrés.}

\item{col.theta}{couleur à appliquer à l'échelle des directions de vents (cf \sQuote{Details}).}

\item{lwd.theta}{épaisseur de l'échelle des directions de vent (cf \sQuote{Details}).}

\item{lty.theta}{type de l'échelle des concentrations (pointillés, continu) (cf \sQuote{Details}).}

\item{centre}{lorsque la rose est ajoutée à un graphique, donne les coordonnées
dans le graphique pré-existant du centre de la rose. Liste avec deux élements
nommés \code{x} et \code{y}.}

\item{pas.rose}{valeur indiquant la précision de la rose (1 -> un point tous les
degrés, 0.5 -> 2 points par degré, etc).}

\item{facteur}{valeur numérique indiquant le grossissement à appliquer à la rose.
Utile uniquement lorsque la rose est ajoutée à un graphique pré-existant pour 
    régler la taille de la rose.}

\item{add}{logical indiquant si la rose doit être ajoutée à un graphique pré-existant.}

\item{\dots}{arguments supplémentaires pour la fonction FUN.}
}
\value{
une liste invisible
\item{breaks}{un vecteur \sQuote{numeric} contenant les limites des secteurs de vents.
	Voir la section\sQuote{Details} pour plus de précisions.}
\item{values}{Vecteur \sQuote{numeric} de longueur égale au nombre de secteurs de vent
	 définis.}
\item{opened}{Un facteur indiquant si les secteurs de vent sont ouverts à gauche ou à droite
	(pour l'instant, seuls les secteurs ouverts à gauche sont pris en compte).}
}
\description{
Cette fonction est un raccourci pour utiliser simultanément les fonctions
\code{\link{compute.rose}} puis \code{\link{plot.rose}}.
}
\seealso{
\code{\link{compute.rose}}, \code{\link{plot.rose}}
}
