% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrGetCampagnes.R
\name{xrGetCampagnes}
\alias{xrGetCampagnes}
\title{Recuperation des campagnes de mesures definies dans une base XR}
\usage{
xrGetCampagnes(
  conn,
  pattern = NULL,
  search.fields = c("NOM_COURT_CM", "LIBELLE"),
  start = NULL,
  end = NULL,
  fields = NULL,
  exact = FALSE
)
}
\arguments{
\item{conn}{une connexion valide telle que retournée par \code{\link{xrConnect}}.}

\item{pattern}{chaînes de caractères utilisées pour la recherche (cf \sQuote{Details}).}

\item{search.fields}{champ de la table dans lesquels \code{pattern} doit être
recherché.}

\item{start}{date initiale de la période à rapatrier (au format \code{\link[base]{POSIXct}} ou character
pouvant être converti en \code{\link[base]{POSIXct}}).}

\item{end}{date de fin de la période à rapatrier (au format \code{\link[base]{POSIXct}} ou character
pouvant être converti en \code{\link[base]{POSIXct}}).}

\item{fields}{vecteurs indiquant les champs de la table à récupérer.
Tous par défaut.}

\item{exact}{booléen indiquant si les mesures à rapatrier doivent être 
exactement identiques à \sQuote{pattern} ou si \sQuote{pattern} doit 
être utilisé dans une expression régulière.}
}
\value{
une data.frame correspondant au contenu de la table 
pour les campagnes trouvées.
}
\description{
La fonction permet de lister les campagnes existantes dans la 
base XR référencée par la connexion \code{conn}.
}
\details{
Il est possible de préciser les champs de recherche,
ou des dates entre lesquelles la station a fonctionné.
}
\seealso{
\code{\link{xrGetContinuousData}}
}
