% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrGetMesures.R
\name{xrGetMesures}
\alias{xrGetMesures}
\title{Recuperation des mesures definies dans une base XR}
\usage{
xrGetMesures(conn, pattern = NULL, search.fields = c("IDENTIFIANT",
  "NOM_COURT_MES"), campagnes = NULL, reseaux = NULL, stations = NULL,
  polluants = NULL, fields = NULL, collapse = c("AND", "OR"),
  exact = FALSE)
}
\arguments{
\item{conn}{une connexion valide telle que retournée par \code{\link{xrConnect}}.}

\item{pattern}{chaînes de caractères utilisées pour la recherche (cf \sQuote{Details}).}

\item{search.fields}{champ de la table dans lesquels \code{pattern} doit être
recherché.}

\item{campagnes}{chaînes de caractères correspondant aux campagnes à rapatrier
    (optionnel) (utilisé via la fonction\code{\link{xrGetCampagnes}}).
 si c'est un vecteur, est directement utilisé comme pattern pour la fonction
 xrGetCampagnes Si c'est une liste, les éléments doivent être nommés. Chaque
 élément est alors utilisé comme argument pour la fonction xrGetCampagnes
  pattern doit alors être précisé :

 \code{... list(pattern='POITIERS_003', search.fields='NOM_COURT_CM') ...}}

\item{reseaux}{chaînes de caractères correspondant aux réseaux à rapatrier
    (optionnel) (utilisé via la fonction\code{\link{xrGetReseaux}}).
 si c'est un vecteur, est directement utilisé comme pattern pour la fonction
 xrGetResaux. Si c'est une liste, les éléments doivent être nommés. Chaque
 élément est alors utilisé comme argument pour la fonction xrGetReseaux.
  pattern doit alors est précisé :

 \code{... list(pattern='R-O3', search.fields='NOM_COURT_RES') ...}}

\item{stations}{chaînes de caractères correspondant aux stations à rapatrier
    (optionnel) (utilisé via la fonction\code{\link{xrGetStations}}).
 si c'est un vecteur, est directement utilisé comme pattern pour la fonction
 xrGetStations. Si C'est une liste, les éléments doivent être nommés. Chaque
 élément est alors utilisé comme argument pour la fonction xrGetStations.
 pattern doit alors être précisé :

 \code{... list(pattern = 'VERDUN', search.fields = 'IDENTIFIANT') ...}}

\item{polluants}{chaînes de caractères correspondant aux polluants à rapatrier
    (optionnel) (utilisé via la fonction\code{\link{xrGetPolluants}}).
 si c'est un vecteur, est directement utilisé comme pattern pour la fonction
 xrGetPolluants. Si C'est une liste, les éléments doivent être nommés. Chaque
 élément est alors utilisé comme argument pour la fonction xrGetPolluants.
 pattern doit alors être précisé :

 \code{... list(pattern = 'O3', search.fields = 'CCHIM') ...}}

\item{fields}{vecteurs indiquant les champs de la table à récupérer.
Tous par défaut.}

\item{collapse}{conjonction à appliquer entre les différents critères de recherche
indiqués.}

\item{exact}{booléen indiquant si les mesures à rapatrier doivent être 
exactement identiques à \sQuote{pattern} ou si \sQuote{pattern} doit 
être utilisé dans une expression régulière.}
}
\value{
une data.frame correspondant au contenu de la table 
pour les mesures trouvées.
}
\description{
La fonction permet de lister les mesures existantes dans la 
base XR référencée par la connexion \code{conn}.
}
\seealso{
\code{\link{xrGetContinuousData}}
}

