% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrGetTypesAnalyseurs.R
\name{xrGetTypesAnalyseurs}
\alias{xrGetTypesAnalyseurs}
\title{Pour chaque mesure, liste les installations des analyseurs.}
\usage{
xrGetTypesAnalyseurs(conn, x, debut, fin)
}
\arguments{
\item{conn}{Connexion à la base XR.}

\item{x}{Une data.frame telle que retournée par
\code{\link[Qair]{xrGetMesures}} avec au moins les colonnes 'NOPOL', 
'NOM_COURT_MES'.}

\item{debut}{POSIXct indiquant la date de début de la période 
pour laquelle on souhaite avoir la liste des analyseurs.}

\item{fin}{POSIXct indiquant la date de fin de la période 
pour laquelle on souhaite avoir la liste des analyseurs.}
}
\value{
Retourne un  TimeInstantDataFrame dont chaque élément contient,
 pour une des mesures demandées, la liste des analyseurs utilisés et la
 date d'installation de l'analyseur.
}
\description{
Pour chaque mesure, liste les installations des analyseurs.
}
