% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agolgets.R
\name{agolGetData}
\alias{agolGetData}
\title{Fonction qui se connecte à ArcGIS online, sur les flux ASQAAs}
\usage{
agolGetData(aasqa, flux, where = "1=1", outFields = "*", getGeometry = TRUE)
}
\arguments{
\item{aasqa}{identifiant 'codé' de l'AASQA correspondant à ses URLs ArcGIS.
Le code en question est la suite de caractères sans signification évidente
qui apparaît systématiquement dans les URLs des flux WFS. Par exemple
'AQwB7zFo5jNpvM0X' dans :
"https://dservices1.arcgis.com/AQwB7zFo5jNpvM0X/arcgis/services/ind_nouvelle_aquitaine/WFSSe..."}

\item{flux}{le nom du flux à interroger.
Il ne s'agit pas exactement du nom du flux mais de celui de la FeatureLayer
qui peut être trouvée sur la page suivante dans l'encadré "URL de requête".
Il s'agit de la partie entre "services/" et "/FeatureServer" (donc
'ind_nouvelle_aquitaine' dans l'exemple).
https://data-atmo-na.opendata.arcgis.com/datasets/ind-nouvelle-aquitaine/geoservice.
Si les FeatureLayer et les flux WFS ont été nommés de la même façon, il 
s'agit tout simplement du nom du flux.}

\item{where}{une requête à faire sur la table du flux. Ça fonctionne comme
une requête SQL.}

\item{outFields}{Vecteurs de chaînes de caractères indiquant les champs à
récupérer. La valeur par défaut '*' signifie 'tous'.}

\item{getGeometry}{Si TRUE, un Spatial*DataFrame est retourné, sinon une data.frame.}
}
\description{
Cette fonction permet de faire une requête sur le flux et donc de n'en 
rappatrier qu'une partie (au lieu du flux complet). Sur des gros flux,
cela permet un gain de temps significatif quand seules quelques valeurs sont
recherchées.
}
\examples{
ingeom <- agolGetData(aasqa='AQwB7zFo5jNpvM0X', flux='ind_nouvelle_aquitaine',
                       where="date\%20IN\%20('2018-09-17','2018-09-18','2018-09-19')")
indice <- agolGetData(aasqa='AQwB7zFo5jNpvM0X', flux='ind_nouvelle_aquitaine',
                       outFields=c('lib_zone', 'date', 'valeur'),
                       where="date\%20IN\%20('2018-09-17','2018-09-18','2018-09-19')",
                       getGeometry=FALSE)
}
