% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agolgets.R
\name{agolListFields}
\alias{agolListFields}
\title{Fonction pour lister les champs d'un flux}
\usage{
agolListFields(aasqa, flux)
}
\arguments{
\item{aasqa}{identifiant 'codé' de l'AASQA correspondant à ses URLs ArcGIS.
Le code en question est la suite de caractères sans signification évidente
qui apparaît systématiquement dans les URLs des flux WFS. Par exemple
'AQwB7zFo5jNpvM0X' dans :
"https://dservices1.arcgis.com/AQwB7zFo5jNpvM0X/arcgis/services/ind_nouvelle_aquitaine/WFSSe..."}

\item{flux}{le nom du flux à interroger ('ind_nouvelle_aquitaine' par exemple).
Contrairement à celui utilisé pour la fonction agolGetData, il s'agit effectivement
du nom du flux tel qu'il apparaît dans les URLs des flux WFS. Par exemple
'ind_nouvelle_aquitaine' dans :
"https://dservices1.arcgis.com/AQwB7zFo5jNpvM0X/arcgis/services/ind_nouvelle_aquitaine/WFSSe..."}
}
\description{
Fonction pour lister les champs d'un flux
}
\examples{
agolListFields(aasqa='AQwB7zFo5jNpvM0X', flux='ind_nouvelle_aquitaine')
}
