% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/reglementation.R
\name{check.na.consecutifs}
\alias{check.na.consecutifs}
\title{Test la validite d'un vecteur par rapport a nombre de NA consecutifs}
\usage{
check.na.consecutifs(x, na.consecutif.max = 720)
}
\arguments{
\item{x}{vecteur à tester}

\item{na.consecutif.max}{entier représentant le nombre maximal
de valeurs NA consécutive autorisé}
}
\value{
TRUE or FALSE see \sQuote{Description}.
}
\description{
Le vecteur est considéré valide si le nombre de NA consécutifs
qui le constitue est inférieur ou égal à \code{na.consecutif.max}.
Si le vecteur est valide, la valeur retournée est TRUE, FALSE sinon.
}

