% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbListFields.R
\name{dbListFields}
\alias{dbListFields}
\alias{dbListFields,RODBC,character-method}
\alias{dbDisconnect}
\alias{dbDisconnect,RODBC-method}
\title{extends dbListFields for RODBC objects}
\usage{
dbListFields(conn, name, ...)

\S4method{dbListFields}{RODBC,character}(conn, name, ...)

dbDisconnect(conn, ...)

\S4method{dbDisconnect}{RODBC}(conn, ...)
}
\arguments{
\item{conn}{A \linkS4class{DBIConnection} object, as returned by
\code{\link[DBI:dbConnect]{dbConnect()}}.}

\item{name}{a character string with the name of the remote table.}

\item{\dots}{pour compatibiltié avec la méthode générique}
}
\description{
extends dbListFields for RODBC objects
}
\seealso{
\code{\link[DBI]{dbListFields}}, \code{\link[RODBC]{sqlColumns}}
}
