% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dbListFields.R
\docType{methods}
\name{dbListFields}
\alias{dbDisconnect}
\alias{dbDisconnect,RODBC-method}
\alias{dbListFields}
\alias{dbListFields,RODBC,character-method}
\title{extends dbListFields for RODBC objects}
\usage{
dbListFields(conn, name, ...)

\S4method{dbListFields}{RODBC,character}(conn, name, ...)

dbDisconnect(conn, ...)

\S4method{dbDisconnect}{RODBC}(conn, ...)
}
\arguments{
\item{conn}{A \code{\linkS4class{DBIConnection}} object, as produced by
\code{\link{dbConnect}}.}

\item{name}{a character string with the name of the remote table.}

\item{...}{Other parameters passed on to methods.}

\item{\dots}{pour compatibiltié avec la méthode générique}
}
\description{
extends dbListFields for RODBC objects
}
\seealso{
\code{\link[DBI]{dbListFields}}, \code{\link[RODBC]{sqlColumns}}
}

