\name{Qair-package}
\alias{Qair-package}
\docType{package}
\title{
Traitement de donnees, domaine de la qualite de l'air
}
\description{
  Ce package definit principalement deux types d'objet qui permettent, via les fonctions definies, de traiter les donnees de qualite de l'air de maniere temporelle et par rapport a des donnees de vent.
  Les fonctions permettent de transformer des donnees d'une base temporelle a une autre, de calculer des profils (journaliers, hebdomadaires, mensuels, annuels) et des roses de pollution.
}
\details{
\tabular{ll}{
Package: \tab Qair\cr
Type: \tab Package\cr
Version: \tab 1.0-1\cr
Date: \tab 2012-02-02\cr
License: \tab GPL\cr
}
}
\author{
Responsable : Vladislav Navel <vladislav.navel@atmopc.org>
}
\seealso{ \code{\link{Qair}} }
\keyword{ package }
