\docType{data}
\name{qh}
\alias{d}
\alias{h}
\alias{m}
\alias{qh}
\alias{y}
\title{definition d'un raccourci pour les heures et les quart-d'heures}
\format{Formal class 'POSIXctp' [package "timetools"] with 2 slots}
\usage{
  qh

  h

  d

  m

  y
}
\description{
  Pour faciliter la manipulation des heures et des
  quart-heures et permettre une utilisation comme c'est le
  cas pour les mois les années etc. dans
  \code{\link[lubridate:seconds]{lubridate}}.
}
\seealso{
  \code{\link[lubridate]{seconds}}
}
\keyword{datasets}

