% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{qh}
\alias{qh}
\alias{h}
\alias{d}
\alias{m}
\alias{y}
\title{definition d'un raccourci pour les heures et les quart-d'heures}
\format{An object of class \code{POSIXctp} of length 1.}
\usage{
qh

h

d

m

y
}
\description{
Pour faciliter la manipulation des heures et des quart-heures
et permettre une utilisation comme c'est le cas pour les mois
les années etc. dans \code{\link[lubridate:seconds]{lubridate}}.
}
\seealso{
\code{\link[lubridate]{seconds}}
}
\keyword{datasets}
