\docType{data}
\name{qh}
\alias{h}
\alias{qh}
\title{definition d'un raccourci pour les heures et les quart-d'heures}
\format{Classes ‘period’ and 'data.frame':	1 obs. of  6 variables:
 $ year  : num 0
 $ month : num 0
 $ day   : num 0
 $ hour  : num 0
 $ minute: num 15
 $ second: num 0}
\usage{
  qh

  h
}
\description{
  Pour faciliter la manipulation des heures et des
  quart-heures et permettre une utilisation comme c'est le
  cas pour les mois les années etc. dans
  \code{\link[lubridate:seconds]{lubridate}}.
}
\seealso{
  \code{\link[lubridate]{seconds}}
}
\keyword{dataset}

