\docType{data}
\name{typologies}
\alias{typologies}
\title{liste les differentes typologie de site possible}
\format{chr [1:6] "industriel" "trafic" "urbain" "périurbain" "rural régional" "rural national"}
\usage{
  typologies
}
\description{
  liste les differentes typologie de site possible
}
\keyword{datasets}

