% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seuils.R
\docType{data}
\name{typologies}
\alias{typologies}
\title{liste les differentes typologie de site possible}
\format{\preformatted{ chr [1:6] "industriel" "trafic" "urbain" "périurbain" "rural régional" "rural national"
}}
\usage{
typologies
}
\description{
liste les differentes typologie de site possible
}
\keyword{datasets}

