% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuils.R
\docType{data}
\name{typologies}
\alias{typologies}
\title{liste les differentes typologie de site possible}
\format{An object of class \code{character} of length 6.}
\usage{
typologies
}
\description{
liste les differentes typologie de site possible
}
\keyword{datasets}

