% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xrGetMethodesPrelevement.R
\name{xrGetMethodesPrelevement}
\alias{xrGetMethodesPrelevement}
\title{Recuperation des methodes de prelevement manuels definies dans une base XR}
\usage{
xrGetMethodesPrelevement(conn, pattern = NULL, search.fields = c("LIBELLE"),
  fields = NULL, collapse = c("AND", "OR"), exact = FALSE)
}
\arguments{
\item{conn}{une connexion valide telle que retournée par \code{\link{xrConnect}}.}

\item{pattern}{chaînes de caractères utilisées pour la recherche (cf \sQuote{Details}).}

\item{search.fields}{champ de la table dans lesquels \code{pattern} doit être
recherché.}

\item{fields}{vecteurs indiquant les champs de la table à récupérer.
    Tous par défaut.}

\item{collapse}{conjonction à appliquer entre les différents critères de recherche
indiqués.}

\item{exact}{booléen indiquant si les mesures à rapatrier doivent être
    exactement identiques à \sQuote{pattern} ou si \sQuote{pattern} doit
    être utilisé dans une expression régulière.}
}
\value{
une data.frame correspondant au contenu de la table
	pour les méthodes trouvées.
}
\description{
La fonction permet de lister les méthodes de prélèvement existantes dans la
base XR référencée par la connexion \code{conn}.
}
\seealso{
\code{\link{xrGetManualData}}, \code{\link{xrGetSitesPrelevement}}
}

