% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrGetPolluants.R
\name{xrGetPolluants}
\alias{xrGetPolluants}
\title{Recuperation des polluants definis dans une base XR}
\usage{
xrGetPolluants(conn, pattern = NULL, search.fields = c("NOPOL", "CCHIM",
  "NCON"), fields = NULL, exact = FALSE)
}
\arguments{
\item{conn}{une connexion valide telle que retournée par \code{\link{xrConnect}}.}

\item{pattern}{chaînes de caractères utilisées pour la recherche (cf \sQuote{Details}).}

\item{search.fields}{champ de la table dans lesquels \code{pattern} doit être
recherché.}

\item{fields}{vecteurs indiquant les champs de la table à récupérer.
Tous par défaut.}

\item{exact}{booléen indiquant si les mesures à rapatrier doivent être 
exactement identiques à \sQuote{pattern} ou si \sQuote{pattern} doit 
être utilisé dans une expression régulière.}
}
\value{
une data.frame correspondant au contenu de la table 
pour les polluants trouvés.
}
\description{
La fonction permet de lister les polluants existants dans la 
base XR référencée par la connexion \code{conn}.
}
\seealso{
\code{\link{xrGetContinuousData}}
}

